+++ dotCMS/src/main/java/com/dotcms/experiments/business/web/ExperimentWebAPI.java
@@ -1,9 +1,12 @@
 package com.dotcms.experiments.business.web;
 
+import com.dotcms.experiments.business.ConfigExperimentUtil;
 import com.dotcms.experiments.model.Experiment;
+import com.dotmarketing.beans.Host;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotSecurityException;
 import java.util.List;
+import java.util.Optional;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -43,4 +46,14 @@ public interface ExperimentWebAPI {
      */
     SelectedExperiments isUserIncluded(final HttpServletRequest request,
             final HttpServletResponse response, final List<String> idsToExclude) throws DotDataException, DotSecurityException ;
+
+
+    /**
+     * Return the HTML/JS Code needed to support Experiments into the Browser
+     * @param host
+     * @param request
+     * @return
+     */
+    public Optional<String> getCode(final Host host, final HttpServletRequest request);
+
 }
