+++ dotCMS/src/main/java/com/dotcms/experiments/business/web/ExperimentWebAPIImpl.java
@@ -1,10 +1,13 @@
 package com.dotcms.experiments.business.web;
 
 import static com.dotcms.util.CollectionsUtils.list;
+import static com.dotmarketing.util.FileUtil.getFileContentFromResourceContext;
 
+import com.dotcms.experiments.business.ConfigExperimentUtil;
 import com.dotcms.experiments.model.Experiment;
 import com.dotcms.experiments.model.ExperimentVariant;
 import com.dotcms.experiments.model.TrafficProportion;
+import com.dotmarketing.beans.Host;
 import com.dotmarketing.business.APILocator;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotRuntimeException;
@@ -12,21 +15,17 @@ import com.dotmarketing.exception.DotSecurityException;
 import com.dotmarketing.portlets.contentlet.model.Contentlet;
 import com.dotmarketing.portlets.htmlpageasset.model.HTMLPageAsset;
 import com.dotmarketing.portlets.rules.model.Rule;
-import com.dotmarketing.util.Config;
+import com.dotmarketing.util.Logger;
 import com.dotmarketing.util.UtilMethods;
 import com.liferay.util.StringPool;
-import java.time.Duration;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
+import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 
 import java.util.Collections;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import net.bytebuddy.utility.RandomString;
@@ -166,5 +165,84 @@ public class ExperimentWebAPIImpl implements ExperimentWebAPI {
 
         rule.get().checkValid();
         return rule.get().evaluate(request, response);
+    }
+
+    /**
+     * Return the Experiment Js Code to inject when there are not any Experiment Running
+     * @return
+     */
+    private String getNotExperimentRunningJSCode() {
+        return "<SCRIPT>localStorage.removeItem('experiment_data');</SCRIPT>";
+    }
+
+    /**
+     * Return the Experiment Js Code to inject when there are any Experiment Running
+     *
+     * @param host Host to use the {@link AnalyticsApp}
+     * @param request To get the Domain name
+     * @return
+     */
+    private String getJSCode(final Host host, final HttpServletRequest request) {
+
+        try {
+            final String analyticsKey = ConfigExperimentUtil.INSTANCE.getAnalyticsKey(host);
+            final String jsJitsuCode =  getFileContentFromResourceContext("experiment/html/experiment_head.html")
+                    .replaceAll("\\$\\{jitsu_key}", analyticsKey)
+                    .replaceAll("\\$\\{site}", getLocalServerName(request));
+
+            final String runningExperimentsId = APILocator.getExperimentsAPI().getRunningExperiments().stream()
+                    .map(experiment -> "'" + experiment.id().get() + "'")
+                    .collect(Collectors.joining(","));
+
+            final String shouldBeInExperimentCalled =  getFileContentFromResourceContext("experiment/js/init_script.js")
+                    .replaceAll("\\$\\{running_experiments_list}", runningExperimentsId);
+
+            return jsJitsuCode + "\n<SCRIPT>" + shouldBeInExperimentCalled + "</SCRIPT>";
+        } catch (IOException | DotDataException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private String getLocalServerName(HttpServletRequest request) {
+        return request.getLocalName() + ":" + request.getLocalPort();
+    }
+
+    /**
+     * Return the HTML/JS Code needed to support Experiments into the Browser, this code is
+     * generated according the follows rules:
+     *
+     * - If the {@link ConfigExperimentUtil#isExperimentEnabled()} is disabled then a
+     * {@link Optional#empty()} is returned.
+     * - If the {@link ConfigExperimentUtil#isExperimentAutoJsInjection()} is disabled then a
+     * {@link Optional#empty()} is returned.
+     * - If both {@link ConfigExperimentUtil#isExperimentEnabled()} and {@link ConfigExperimentUtil#isExperimentAutoJsInjection()}
+     * are TRUE but we don't have any {@link com.dotcms.experiments.model.Experiment} RUNNING then
+     * the generated code is just to remove the localStorage object if it exists.
+     * - If both {@link ConfigExperimentUtil#isExperimentEnabled()} and {@link ConfigExperimentUtil#isExperimentAutoJsInjection()}
+     * are TRUE and also it is at least one  {@link com.dotcms.experiments.model.Experiment} RUNNING then
+     * the generated code do the Follow:
+     *      - Hit the Endpoint to know if the USer should go into the Experiment.
+     *      - Create the localStorage Object to keep all the Experiment's data needed.
+     *      - Redirect to a Page's Variant version if it is needed according to the Variant into what the user was assigned.
+     *      - Clean the localStorage when a Experiment is sttoped but still exists at least one more RUNNING.
+     *
+     * @param host
+     * @param request
+     * @return
+     */
+    @Override
+    public Optional<String> getCode(final Host host, final HttpServletRequest request) {
+        if (ConfigExperimentUtil.INSTANCE.isExperimentEnabled()) {
+
+            try {
+                final String code = APILocator.getExperimentsAPI().isAnyExperimentRunning() ?
+                        getJSCode(host, request) : getNotExperimentRunningJSCode();
+                return Optional.of(code);
+            } catch (DotDataException e) {
+                Logger.error(ExperimentWebAPIImpl.class, "It is not possible to generate the Experiment JS COde:" + e.getMessage());
+            }
+        }
+
+        return Optional.empty();
     }
 }
