+++ dotCMS/src/main/java/com/dotcms/rendering/velocity/viewtools/experiment/ExperimentCodeInjecterViewTool.java
@@ -1,9 +1,8 @@
 package com.dotcms.rendering.velocity.viewtools.experiment;
 
 import com.dotcms.api.web.HttpServletRequestThreadLocal;
-import com.dotcms.experiments.business.ExperimentCodeGenerator;
+import com.dotcms.experiments.business.web.ExperimentWebAPI;
 import com.dotmarketing.beans.Host;
-import com.dotmarketing.business.APILocator;
 import com.dotmarketing.business.web.WebAPILocator;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotRuntimeException;
@@ -17,6 +16,11 @@ import org.apache.velocity.tools.view.tools.ViewTool;
  * ViewTool to inject the JS/HTML Experiment Code
  */
 public class ExperimentCodeInjecterViewTool implements ViewTool {
+    final ExperimentWebAPI experimentWebAPI;
+
+    public ExperimentCodeInjecterViewTool() {
+        this.experimentWebAPI = WebAPILocator.getExperimentWebAPI();
+    }
 
     @Override
     public void init(Object initData) {
@@ -32,7 +36,7 @@ public class ExperimentCodeInjecterViewTool implements ViewTool {
         try {
             final Host currentHost = WebAPILocator.getHostWebAPI().getCurrentHost();
             final HttpServletRequest request = HttpServletRequestThreadLocal.INSTANCE.getRequest();
-            return ExperimentCodeGenerator.INSTANCE.getCode(currentHost, request).orElse(StringPool.BLANK);
+            return experimentWebAPI.getCode(currentHost, request).orElse(StringPool.BLANK);
         } catch (DotDataException | DotSecurityException e) {
             throw new DotRuntimeException(e);
         }
